+++ src/utils/index.ts
@@ -25,16 +25,16 @@ export const usdConverter = async (weiVal:bigint | undefined) => {
 }
 
 export function formatStringUnits(display: string, decimals: number) {
-  const negative = display.startsWith("-");
+  const negative = display?.startsWith("-");
   if (negative) display = display.slice(1);
 
-  display = display.padStart(decimals, "0");
+  display = display?.padStart(decimals, "0");
 
   let [integer, fraction] = [
-    display.slice(0, display.length - decimals),
-    display.slice(display.length - decimals),
+    display?.slice(0, display.length - decimals),
+    display?.slice(display.length - decimals),
   ];
-  fraction = fraction.replace(/(0+)$/, "");
+  fraction = fraction?.replace(/(0+)$/, "");
   return `${negative ? "-" : ""}${integer || "0"}${
     fraction ? `.${fraction}` : ""
   }`;
